fix(frontend): Web 切り取り後のアップロードで画像が届かない不具合を修正
Web の FormData は File/Blob が必要で、RN 形式 { uri, name, type } では
multer にファイルが渡らなかった。切り取り後の blob URL を fetch して File 化。
multipart の Content-Type 手動指定も削除（boundary 付与のため）。
